Мелкие исправления. Подготовка для публикации на GitHub.

+++ Controllers/MoviesController.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -13,6 +13,9 @@ using SkillsTest.Models;
 
 namespace SkillsTest.Controllers
 {
+    /// <summary>
+    /// Контроллер для работы с каталогом фильмов.
+    /// </summary>
     [Authorize]
     public class MoviesController : Controller
     {
@@ -27,6 +30,15 @@ namespace SkillsTest.Controllers
             _mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
         }
 
+        /// <summary>
+        /// Главная страница. Отображает каталог фильмов.
+        /// </summary>
+        /// <param name="pageSize">Количество фильмов на странице.</param>
+        /// <param name="page">Номер страницы.</param>
+        /// <param name="configuration">
+        /// Объект для доступа к конфигурации приложения в файле appSettings.json. Количество фильмов на странице
+        /// по умолчанию и максимальное хранятся в файле концигурации в секции MoviesSettings.
+        /// </param>
         [AllowAnonymous]
         public async Task<IActionResult> Index(int pageSize, int page, [FromServices] IConfiguration configuration)
         {
@@ -34,19 +46,26 @@ namespace SkillsTest.Controllers
                 defaultPageSize = int.Parse(configuration["MoviesSettings:DefaultPageSize"]),
                 maxPageSize = int.Parse(configuration["MoviesSettings:MaxPageSize"]);
 
+            // Если номер страницы 0 или отрицательное число, то редиректим на первую страницу.
             if (page < 1)
                 return RedirectToAction("Index", new { page = 1, pageSize });
 
+            // Если количество фильмов на странице меньше значения по умолчанию или больше максимального, то редиректим
+            // на указанную страницу с размером страницы по умолчанию.
             if (pageSize < defaultPageSize || pageSize > maxPageSize)
                 return RedirectToAction("Index", new { page, pageSize = defaultPageSize });
 
-            int 
+            int
+                // Общее количество фильмов в каталоге.
                 totalCount = _movieRepository.TotalCount,
+                // Сколько всего страниц при заданном количестве фильмов на страницу.
                 maxPage = (int)Math.Ceiling(totalCount * 1d / pageSize);
 
+            // Если номер текущей страницы больше, чем всего страниц, то редиректим на последнюю страницу.
             if (maxPage != 0 && page > maxPage)
                 return RedirectToAction("Index", new { page = maxPage, pageSize });
 
+            // Определим смещение - сколько фильмов пропустить от начала.
             int offset = page * pageSize - pageSize;
             var movies = await _movieRepository.GetMoviesAsync(pageSize, offset);
             var model = new PaginatedMoviesViewModel
@@ -76,10 +95,12 @@ namespace SkillsTest.Controllers
         [Route("movie/{id:int}/edit")]
         public async Task<IActionResult> Edit(int id)
         {
+            // Если фильма с таким ID не существует, то редиректим на страницу 404.
             var movie = await _movieRepository.GetMovieAsync(id);
             if (movie == null)
                 return NotFound();
 
+            // Если редактирование фильма запрашивает не его создатель, то редиректим на страницу 403.
             if (movie.CreatedBy != User.Identity.Name)
                 return Forbid();
 
@@ -94,6 +115,7 @@ namespace SkillsTest.Controllers
             if (!ModelState.IsValid)
                 return View(model);
 
+            // Небольшой гвард, вдруг в запросе сохранение редактирования придёт неправильный ID.
             var entity = await _movieRepository.GetMovieAsync(model.Id);
             if (entity == null)
             {
@@ -101,23 +123,33 @@ namespace SkillsTest.Controllers
                 return View(model);
             }
 
+            // Если редактирование фильма запрашивает не его создатель, то редиректим на страницу 403.
             if (model.CreatedBy != User.Identity.Name)
                 return Forbid();
 
             var movie = _mapper.Map<Movie>(model);
             if (model.Image != null)
             {
+                // Если пользователь хочет сменить постер фильма, то сохраняем новое изображение в папке 'user-images'.
                 string extension = Path.GetExtension(model.Image.FileName);
                 movie.Poster = await _imageRepository.SaveImageAsync(model.Image.OpenReadStream(), extension);
             }
             else
             {
+                // Иначе оставляем прежний постер.
                 movie.Poster = model.Poster;
             }
 
+            // Сохранить ссылку на прежний постер фильма.
             string oldPoster = entity.Poster;
+
+            // Обновить данные фильма в репозитории.
             _movieRepository.SaveMovie(movie);
-            _imageRepository.DeleteImage(Path.GetFileName(oldPoster));
+
+            // Только после того, как данные фильма были успешно обновлены и если постер был обновлён, то удалить
+            // старый постер.
+            if (model.Image != null)
+                _imageRepository.DeleteImage(Path.GetFileName(oldPoster));
             
             TempData.Add("Message", "Данные фильма успешно обновлены.");
             var updatedModel = _mapper.Map<MovieBindingModel>(movie);
@@ -141,6 +173,8 @@ namespace SkillsTest.Controllers
         [HttpPost]
         public async Task<IActionResult> Add(MovieBindingModel model)
         {
+            // Необходимых встроенных атрибутов валидации 'IFormFile' нет, по-хорошему надо бы написать кастомный,
+            // но ограничимся ручной проверкой.
             if (model.Image == null)
                 ModelState.AddModelError("Image", "Необходимо добавить постер фильма.");
 
@@ -163,12 +197,19 @@ namespace SkillsTest.Controllers
             if (movie == null)
                 return NotFound();
 
+            // Не позволим пользователю, не являющемуся создателем фильма, удалить его.
             if (movie.CreatedBy != User.Identity.Name)
                 return Forbid();
 
+            // Сохранить ссылку на файл изображения постера фильма.
             string posterFilePath = movie.Poster;
+
+            // Удалить фильм из репозитория.
             _movieRepository.DeleteMovie(movie);
+
+            // Если фильм удалён успешно, то удалить и постер.
             _imageRepository.DeleteImage(Path.GetFileName(posterFilePath));
+
             return RedirectToAction("Index", "Movies");
         }
     }
